# lib/api-zod/src/generated/types/createStaffRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateStaffRequest {
  /** @minLength 1 */
  firebaseUid: string;
  /** @minLength 1 */
  displayName: string;
  email?: string;
  roleIds: string[];
  active: boolean;
  photoUrl?: string;
}