# lib/api-zod/src/generated/types/role.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RolePermissionsItem } from './rolePermissionsItem';

export interface Role {
  id: string;
  clubId?: string;
  /** @minLength 1 */
  name: string;
  permissions: RolePermissionsItem[];
  active: boolean;
}